cattegorized componenets and split them into files
