+++ GUI/src/main/kotlin/ui/components/general/AutoSizeText.kt
@@ -1,4 +1,4 @@
-package ui.components
+package ui.components.general
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.LocalTextStyle
