+++ GUI/src/main/kotlin/ui/components/general/InfoIconWithHover.kt
@@ -1,4 +1,4 @@
-package ui.components
+package ui.components.general
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.size
