+++ GUI/src/main/kotlin/ui/components/general/TextTitle.kt
@@ -1,4 +1,4 @@
-package ui.components
+package ui.components.general
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.runtime.Composable
