+++ GUI/src/main/kotlin/ui/components/general/Tooltip.kt
@@ -1,4 +1,4 @@
-package ui.components
+package ui.components.general
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
