+++ GUI/src/main/kotlin/ui/components/settingsScreen/SaveButton.kt
@@ -1,0 +1,35 @@
+package ui.components.settingsScreen
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Button
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import models.AppState
+
+@Composable
+fun RowScope.SaveButton(
+    state: MutableState<AppState>,
+    oldState: MutableState<AppState>,
+) {
+    Button(
+        // fills all available space
+        modifier = Modifier.weight(0.1f).padding(8.dp).fillMaxSize(1f),
+        onClick = {
+            oldState.value = state.value
+            state.value = oldState.value.copy(screen = Screen.SelectVideoScreen)
+        },
+    ) {
+        Image(
+            painter = painterResource("save.svg"),
+            contentDescription = "save",
+            modifier = Modifier.fillMaxSize().alpha(0.8f).padding(4.dp),
+        )
+    }
+}
