+++ GUI/src/main/kotlin/ui/themes/ThemeWrapper.kt
@@ -1,0 +1,14 @@
+package ui.themes
+
+import androidx.compose.material.Surface
+import androidx.compose.runtime.Composable
+
+// sets the default color palette
+@Composable
+fun wrapTheming(content: @Composable () -> Unit) {
+    defaultTheme {
+        Surface(color = defaultBackgroundColor) {
+            content()
+        }
+    }
+}
